+++ src/questionController.js
@@ -1,35 +1,64 @@
 const { getDb } = require('./config/database');
-// const { connectToDatabase } = require('./config/database');
 
 async function createQuestion(questionData) {
-    const db = getDb();
-    const existingAnswer = await db.collection('Questions').findOne({ id: questionData.id });
-    if (existingAnswer) {
-        throw new Error(`La réponse avec l'ID ${questionData.id} existe déjà.`);
+    try {
+        const db = getDb();
+        const existingAnswer = await db.collection('questions').findOne({ id: questionData.id });
+        if (existingAnswer) {
+            throw new Error(`La réponse avec l'ID ${questionData.id} existe déjà.`);
+        }
+        const result = await db.collection('questions').insertOne(questionData);
+        console.log("question ajouté");
+        
+        return result.insertedId;
+    } catch (error) {
+        console.log(error.message)
     }
-    const result = await db.collection('questions').insertOne(questionData);
-    console.log("question ajouté");
-    
-    return result;
 }
 
 async function getQuestionById() {
-    const db = getDb();
-    const question = await db.collection('questions').find({ }).toArray();
-    return question;
+    
+    try {
+        const db = getDb();
+        const question = await db.collection('questions').find({ }).toArray();
+        if (question.length === 0) {
+            throw new Error("aucune question n'est trouvée")
+        }
+        console.log('Question:', question);
+    } catch (error) {
+        console.log(error.message)
+    }
 }
 
 async function updateQuestion(id, updateData) {
-    const db = getDb();
-    const result = await db.collection('questions').updateOne({ id: id }, { $set: updateData });
-    return result;
+    try {
+        const db = getDb();
+        const result = await db.collection('questions').updateOne({ id }, { $set: updateData });
+        const existingAnswer = await db.collection('questions').findOne({ id: id });
+        if (!existingAnswer) {
+            throw new Error(`La question avec l'ID ${id} n'existe pas.`);
+        }
+        console.log('Question est modifiée.');
+        return result;
+    } catch (error) {
+        console.log(error.message)
+    }
 }
 
 
 async function deleteQuestion(id) {
-    const db = getDb();
-    const result = await db.collection('questions').deleteOne({ id: id });
-    return result;
+    try {
+        const db = getDb();
+        const existingAnswer = await db.collection('questions').findOne({ id: id });
+        if (!existingAnswer) {
+            throw new Error(`La question avec l'ID ${id} n'existe pas.`);
+        }
+        const result = await db.collection('questions').deleteOne({ id: id });
+        console.log('Question deleted.');
+        return result;
+    } catch (error) {
+        console.log(error.message)
+    }
 }
 
 module.exports = { createQuestion, getQuestionById, updateQuestion, deleteQuestion };
